+++ src/net/p2p/peer_score.hpp
@@ -25,7 +25,7 @@ public:
     const fly::net::Addr& addr();
     void add_score(uint64 score);
     void sub_score(uint64 score);
-    std::atomic<uint32> m_state{0};
+    std::atomic<uint8> m_state{0};
     
 private:
     uint64 m_score;
